+++ commfuncs.h
@@ -28,7 +28,7 @@
 
 using namespace std;
 
-#define VERSION "v0.932"
+#define VERSION "v0.933"
 
 #define UNKNOWN 0